Update wrangler to CUDA Toolkit 11.8
This started as an investigation of the #10.

The actual pycparser version does not seem to matter too much,
is just some syntax in the header changed in the past years,
so the wrangler code needed an update to not fail.

This change contains fixes to make the auto-generator work on
more modern machines (with newer CPP versions for example), and
also makes it so certain syntax in type definition is parsed
and converted to string correctly.